Extract SdJwtVcJwtProcessor to its own file.
