+++ src/main/kotlin/eu/europa/ec/eudi/sdjwt/vc/SdJwtVcVerifier.kt
@@ -15,25 +15,12 @@
  */
 package eu.europa.ec.eudi.sdjwt.vc
 
-import com.nimbusds.jose.JOSEObjectType
-import com.nimbusds.jose.JWSAlgorithm
-import com.nimbusds.jose.JWSVerifier
-import com.nimbusds.jose.crypto.ECDSAVerifier
-import com.nimbusds.jose.crypto.Ed25519Verifier
-import com.nimbusds.jose.crypto.MACVerifier
-import com.nimbusds.jose.crypto.RSASSAVerifier
 import com.nimbusds.jose.jwk.*
 import com.nimbusds.jose.jwk.source.ImmutableJWKSet
 import com.nimbusds.jose.jwk.source.JWKSource
-import com.nimbusds.jose.proc.BadJOSEException
-import com.nimbusds.jose.proc.DefaultJOSEObjectTypeVerifier
 import com.nimbusds.jose.proc.SecurityContext
 import com.nimbusds.jose.util.X509CertUtils
-import com.nimbusds.jwt.JWTClaimsSet
 import com.nimbusds.jwt.SignedJWT
-import com.nimbusds.jwt.proc.BadJWTException
-import com.nimbusds.jwt.proc.DefaultJWTClaimsVerifier
-import com.nimbusds.jwt.proc.DefaultJWTProcessor
 import eu.europa.ec.eudi.sdjwt.*
 import eu.europa.ec.eudi.sdjwt.vc.SdJwtVcIssuerPublicKeySource.*
 import io.ktor.http.*
@@ -41,7 +28,6 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
 import kotlinx.serialization.json.JsonObject
 import java.security.cert.X509Certificate
-import java.text.ParseException
 
 fun interface X509CertificateTrust {
     suspend fun isTrusted(chain: List<X509Certificate>): Boolean
@@ -278,79 +264,3 @@ internal fun keySource(jwt: SignedJWT): SdJwtVcIssuerPublicKeySource? {
         else -> null
     }
 }
-
-const val SD_JWT_VC_TYPE = "vc+sd-jwt"
-
-/**
- * Custom [DefaultJWTProcessor] implementation with support for [OctetKeyPair].
- */
-private class SdJwtVcJwtProcessor<C : SecurityContext>(
-    private val jwkSource: JWKSource<C>,
-) : DefaultJWTProcessor<C>() {
-    init {
-        jwsTypeVerifier = DefaultJOSEObjectTypeVerifier(JOSEObjectType(SD_JWT_VC_TYPE))
-        jwtClaimsSetVerifier = DefaultJWTClaimsVerifier(
-            JWTClaimsSet.Builder().build(),
-            setOf("iss"),
-        )
-    }
-
-    override fun process(signedJWT: SignedJWT, context: C?): JWTClaimsSet {
-        ensureInitialized()
-
-        jwsTypeVerifier.verify(signedJWT.header.type, context)
-
-        val claimsSet = signedJWT.jwtClaimSet()
-        val jwkSelector = JWKSelector(JWKMatcher.forJWSHeader(signedJWT.header))
-
-        val jwks = jwkSource.get(jwkSelector, context)
-        if (jwks.isNullOrEmpty()) {
-            throw BadJOSEException("Signed JWT rejected: Another algorithm expected, or no matching key(s) found")
-        }
-
-        for (jwk in jwks) {
-            val verifier = jwsVerifierFor(signedJWT.header.algorithm, jwk)
-            if (signedJWT.verify(verifier)) {
-                jwtClaimsSetVerifier.verify(claimsSet, context)
-                return claimsSet
-            }
-        }
-
-        // No more keys to try out
-        throw BadJOSEException("Signed JWT rejected: Invalid signature or no matching verifier(s) found")
-    }
-
-    private fun ensureInitialized() {
-        if (jwsTypeVerifier == null) {
-            throw BadJOSEException("Signed JWT rejected: No JWS header typ (type) verifier is configured")
-        }
-
-        if (jwtClaimsSetVerifier == null) {
-            throw BadJOSEException("Signed JWT rejected: No JWTClaimSet verifier is configured")
-        }
-    }
-}
-
-private fun SignedJWT.jwtClaimSet(): JWTClaimsSet =
-    try {
-        getJWTClaimsSet()
-    } catch (e: ParseException) {
-        // Payload not a JSON object
-        throw BadJWTException(e.message, e)
-    }
-
-private fun jwsVerifierFor(algorithm: JWSAlgorithm, jwk: JWK): JWSVerifier =
-    when (algorithm) {
-        in JWSAlgorithm.Family.HMAC_SHA -> MACVerifier(jwk.expectIs<OctetSequenceKey>())
-        in JWSAlgorithm.Family.RSA -> RSASSAVerifier(jwk.expectIs<RSAKey>())
-        in JWSAlgorithm.Family.EC -> ECDSAVerifier(jwk.expectIs<ECKey>())
-        in JWSAlgorithm.Family.ED -> Ed25519Verifier(jwk.expectIs<OctetKeyPair>())
-        else -> throw BadJOSEException("Unsupported JWS algorithm $algorithm")
-    }
-
-private inline fun <reified T> JWK.expectIs(): T =
-    if (this is T) {
-        this
-    } else {
-        throw BadJOSEException("Expected a JWK of type ${T::class.java.simpleName}")
-    }
